requirements/dev.txt
-r base.txt
pytest-django==4.5.2
pytest-factoryboy==2.5.0
